Allow fetching with no GITHUB_TOKEN
