+++ src/Root.tsx
@@ -1,12 +1,12 @@
 import {useCallback} from 'react';
 import {CalculateMetadataFunction, Composition} from 'remotion';
-import {fetchStargazers} from './fetch';
 import {Main, MainProps, mainSchema} from './Main';
+import {fetchStargazers} from './fetch/fetch-data';
 import {waitForNoInput} from './wait-for-no-input';
 
 const FPS = 30;
 
-export const RemotionRoot = () => {
+export function RemotionVideo() {
 	const calculateMetadata: CalculateMetadataFunction<MainProps> = useCallback(
 		async ({props, abortSignal}) => {
 			await waitForNoInput(abortSignal, 500);
@@ -33,6 +33,7 @@ export const RemotionRoot = () => {
 		<Composition
 			id="main"
 			component={Main}
+			durationInFrames={15 * FPS}
 			fps={FPS}
 			width={960}
 			height={540}
@@ -47,4 +48,4 @@ export const RemotionRoot = () => {
 			}}
 		/>
 	);
-};
+}
